+++ test/cli/output/test_default.py
@@ -12,6 +12,7 @@ from schemathesis.cli.output import default
 from schemathesis.cli.output.default import display_internal_error
 from schemathesis.runner.events import Finished, InternalError
 from schemathesis.runner.serialization import SerializedTestResult
+from schemathesis.constants import USER_AGENT
 
 from ...utils import strip_style_win32